Refactor
- 코드리뷰 반영
- BackPressHandler -> BackPressInterceptor로 네이밍 변경
- 뒤로가기 눌렀을 때 기본으로는 1.5초 체크해서 액티비티를 종료시키지만, 바텀시트 등에서 뒤로가기 눌렀을 때는 hidden 등 옵션을 처리할 수 있도록 수정
- SplashScreen.kt에 필요없는 주석 제거
